requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu130

cd_fvd==0.1.1
decord==0.6.0
diffusers==0.36.0
huggingface_hub==1.10.1
imageio==2.37.2
imageio-ffmpeg==0.6.0
matplotlib==3.10.8
moviepy==2.2.1
numpy==2.4.1
opencv_python==4.13.0.90
Pillow==11.3.0
timm==1.0.24
torch==2.10.0
torchmetrics[image]==1.9.0
torchvision==0.25.0
tqdm==4.67.1
wandb==0.24.1
